+++ trabalho-pratico/include/entidades/user.h
@@ -12,7 +12,7 @@
 typedef struct user User;
 
 // Função para criar uma estrutura da entidade utilizador parametrizada.
-User* createUser(char* username, char* email, char* first_name, char* last_name, char* birth_date , char* country, char* subscription_type, int* liked_musics_id, int num_liked_musics);
+User* createUser(char* username, char* email, char* first_name, char* last_name, char* birth_date , char* country, char* subscription_type, long int* liked_musics_id, int num_liked_musics);
 
 // Função para libertar a memória de uma entidade do tipo utilizador.
 void freeUser(User* user);
